Lexer

Description

Documentation

data Token

Type representing a token.

Constructors

Keyword KeywordT
A keyword of the language

Operator OperatorT
An operator

Id String
The name of an identifier

Symbol SymbolT
A symbol

Number Integer
An integer number

String String
A string

Bool Bool
A boolean value

Nil
The empty list

data SymbolT

Type representing symbols (parenthesis, comma...).

Constructors

LPAREN
Left parenthesis

RPAREN
Right parenthesis

EQUALS
Equality symbol

PLUS
Sum symbol

MINUS
Difference symbol

TIMES
Multiplication symbol

DIVISION
Division symbol

COMMA
Comma symbol, used to separate actual parameters

DOLLAR
Dollar symbol, meaning end of input

data OperatorT

Type representing operators.

Constructors

EQ
Equality test

LEQ
Less or equal test

CAR
Operator returning the list head

CDR
Operator returning the list tail

CONS
Constructor of a list

ATOM
Tells whether its argument is a value

data KeywordT

Type representing keywords.

Constructors

LET
Keyword used to declare a binding block

LETREC
Keyword used to declare a recursive binding block

IN
Keyword indicating the body of a function

END
Keyword indicating the end of a function

AND
Used to concatenate bindings

IF
Part of the if-then-else branch

THEN
Part of the if-then-else branch

ELSE
Part of the if-theb-else branch

LAMBDA
Keyword used to declare a function

lexi :: String -> [Token]

Main function for the lexical analysis. Scans string representing the source code of the program and converts it into a list of tokens.