Samples

Description

Documentation

sample_prog_1 :: String

Sample program which computes the factorial for any element of a list and store the results into a new list.

sample_prog_2 :: String

Sample program which builds a list.